Refactoring of online and offline signature verification
