+++ powerauth-java-server/src/main/java/io/getlime/security/powerauth/app/server/service/behavior/tasks/v3/RecoveryServiceBehavior.java
@@ -576,7 +576,7 @@ public class RecoveryServiceBehavior {
     /**
      * Update recovery configuration.
      * @param request Update recovery configuration request.
-     * @param keyConversion Key conversion utility.
+     * @param keyConversion Key convertor.
      * @return Update recovery configuration response.
      * @throws GenericServiceException In case of any error.
      */
