+++ powerauth-java-server/src/main/java/io/getlime/security/powerauth/app/server/service/behavior/tasks/v3/VaultUnlockServiceBehavior.java
@@ -200,7 +200,7 @@ public class VaultUnlockServiceBehavior {
             additionalInfo.getEntry().add(entry);
 
             // Verify the signature
-            VerifySignatureResponse signatureResponse = behavior.getSignatureServiceBehavior().verifySignature(activationId, signatureType,
+            VerifySignatureResponse signatureResponse = behavior.getOnlineSignatureServiceBehavior().verifySignature(activationId, signatureType,
                     signature, additionalInfo, signedData, applicationKey, null, keyConversion);
 
             VaultUnlockResponsePayload responsePayload = new VaultUnlockResponsePayload();
